// ReactInputy/src/api/generated/models/AddBloodDonatorDTO.ts
/* tslint:disable */
/* eslint-disable */
/**
 * BloodBank
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: Wersja 1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
import {
    AddUserDTO,
    AddUserDTOFromJSON,
    AddUserDTOFromJSONTyped,
    AddUserDTOToJSON,
    BloodTypeNameDTO,
    BloodTypeNameDTOFromJSON,
    BloodTypeNameDTOFromJSONTyped,
    BloodTypeNameDTOToJSON,
} from './';

/**
 * 
 * @export
 * @interface AddBloodDonatorDTO
 */
export interface AddBloodDonatorDTO {
    /**
     * 
     * @type {string}
     * @memberof AddBloodDonatorDTO
     */
    pesel?: string | null;
    /**
     * 
     * @type {string}
     * @memberof AddBloodDonatorDTO
     */
    homeAdress?: string | null;
    /**
     * 
     * @type {string}
     * @memberof AddBloodDonatorDTO
     */
    phoneNumber?: string | null;
    /**
     * 
     * @type {BloodTypeNameDTO}
     * @memberof AddBloodDonatorDTO
     */
    bloodType?: BloodTypeNameDTO;
    /**
     * 
     * @type {AddUserDTO}
     * @memberof AddBloodDonatorDTO
     */
    user?: AddUserDTO;
}

export function AddBloodDonatorDTOFromJSON(json: any): AddBloodDonatorDTO {
    return AddBloodDonatorDTOFromJSONTyped(json, false);
}

export function AddBloodDonatorDTOFromJSONTyped(json: any, ignoreDiscriminator: boolean): AddBloodDonatorDTO {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'pesel': !exists(json, 'pesel') ? undefined : json['pesel'],
        'homeAdress': !exists(json, 'homeAdress') ? undefined : json['homeAdress'],
        'phoneNumber': !exists(json, 'phoneNumber') ? undefined : json['phoneNumber'],
        'bloodType': !exists(json, 'bloodType') ? undefined : BloodTypeNameDTOFromJSON(json['bloodType']),
        'user': !exists(json, 'user') ? undefined : AddUserDTOFromJSON(json['user']),
    };
}

export function AddBloodDonatorDTOToJSON(value?: AddBloodDonatorDTO | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'pesel': value.pesel,
        'homeAdress': value.homeAdress,
        'phoneNumber': value.phoneNumber,
        'bloodType': BloodTypeNameDTOToJSON(value.bloodType),
        'user': AddUserDTOToJSON(value.user),
    };
}