// ReactInputy/src/api/generated/models/AddUserDTO.ts
/* tslint:disable */
/* eslint-disable */
/**
 * BloodBank
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: Wersja 1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface AddUserDTO
 */
export interface AddUserDTO {
    /**
     * 
     * @type {string}
     * @memberof AddUserDTO
     */
    email?: string | null;
    /**
     * 
     * @type {string}
     * @memberof AddUserDTO
     */
    firstname?: string | null;
    /**
     * 
     * @type {string}
     * @memberof AddUserDTO
     */
    surname?: string | null;
}

export function AddUserDTOFromJSON(json: any): AddUserDTO {
    return AddUserDTOFromJSONTyped(json, false);
}

export function AddUserDTOFromJSONTyped(json: any, ignoreDiscriminator: boolean): AddUserDTO {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'email': !exists(json, 'email') ? undefined : json['email'],
        'firstname': !exists(json, 'firstname') ? undefined : json['firstname'],
        'surname': !exists(json, 'surname') ? undefined : json['surname'],
    };
}

export function AddUserDTOToJSON(value?: AddUserDTO | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'email': value.email,
        'firstname': value.firstname,
        'surname': value.surname,
    };
}