// ReactInputy/src/api/generated/models/BloodTypeNameDTO.ts
/* tslint:disable */
/* eslint-disable */
/**
 * BloodBank
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: Wersja 1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface BloodTypeNameDTO
 */
export interface BloodTypeNameDTO {
    /**
     * 
     * @type {string}
     * @memberof BloodTypeNameDTO
     */
    bloodTypeName?: string | null;
}

export function BloodTypeNameDTOFromJSON(json: any): BloodTypeNameDTO {
    return BloodTypeNameDTOFromJSONTyped(json, false);
}

export function BloodTypeNameDTOFromJSONTyped(json: any, ignoreDiscriminator: boolean): BloodTypeNameDTO {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'bloodTypeName': !exists(json, 'bloodTypeName') ? undefined : json['bloodTypeName'],
    };
}

export function BloodTypeNameDTOToJSON(value?: BloodTypeNameDTO | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'bloodTypeName': value.bloodTypeName,
    };
}